i32 is faster than smaller types

+++ src/san.rs
@@ -153,8 +153,8 @@ impl Error for SanError {
 pub enum San {
     Normal {
         role: Role,
-        file: Option<i8>,
-        rank: Option<i8>,
+        file: Option<i32>,
+        rank: Option<i32>,
         capture: bool,
         to: Square,
         promotion: Option<Role>,
@@ -164,17 +164,17 @@ pub enum San {
     Null,
 }
 
-fn rank_from_char(ch: u8) -> Option<i8> {
+fn rank_from_char(ch: u8) -> Option<i32> {
     if b'1' <= ch && ch <= b'8' {
-        Some((ch - b'1') as i8)
+        Some(i32::from(ch - b'1'))
     } else {
         None
     }
 }
 
-fn file_from_char(ch: u8) -> Option<i8> {
+fn file_from_char(ch: u8) -> Option<i32> {
     if b'a' <= ch && ch <= b'h' {
-        Some((ch - b'a') as i8)
+        Some(i32::from(ch - b'a'))
     } else {
         None
     }
@@ -222,13 +222,13 @@ impl San {
             };
 
             let (file, next) = if b'a' <= next && next <= b'h' {
-                (Some((next - b'a') as i8), chars.next().ok_or(())?)
+                (Some(i32::from(next - b'a')), chars.next().ok_or(())?)
             } else {
                 (None, next)
             };
 
             let (rank, next) = if b'1' <= next && next <= b'8' {
-                (Some((next - b'1') as i8), chars.next())
+                (Some(i32::from(next - b'1')), chars.next())
             } else {
                 (None, Some(next))
             };
